requirements.txt
fastapi==0.105.0
uvicorn==0.24.0.post1
slack_sdk
pydantic==1.10.13
python-dotenv==1.0.1
psycopg==3.1.17
pandas
python-multipart
tabulate
tabulate[widechars] 
matplotlib